+++ lib/src/reactive_builder.dart
@@ -4,12 +4,14 @@ import 'reactive_model.dart';
 class ReactiveBuilder<T extends ReactiveModel> extends StatefulWidget {
   final T model;
   final Widget Function(T model) builder;
-  final List<String>? fields; // optional field filter
+  final List<Symbol>? fields;
+  final bool Function(T oldModel, T newModel)? shouldRebuild;
 
   const ReactiveBuilder({
     required this.model,
     required this.builder,
     this.fields,
+    this.shouldRebuild,
     super.key,
   });
 
@@ -19,7 +21,13 @@ class ReactiveBuilder<T extends ReactiveModel> extends StatefulWidget {
 
 class ReactiveBuilderState<T extends ReactiveModel>
     extends State<ReactiveBuilder<T>> {
-  void _onChange() => setState(() {});
+  void _onChange() {
+    if (!mounted) return;
+    if (widget.shouldRebuild != null) {
+      if (!widget.shouldRebuild!(widget.model, widget.model)) return;
+    }
+    setState(() {});
+  }
 
   @override
   void initState() {
@@ -48,6 +56,58 @@ class ReactiveBuilderState<T extends ReactiveModel>
   @override
   Widget build(BuildContext context) => widget.builder(widget.model);
 }
+
+///Version 0.0.7
+// import 'package:flutter/material.dart';
+// import 'reactive_model.dart';
+//
+// class ReactiveBuilder<T extends ReactiveModel> extends StatefulWidget {
+//   final T model;
+//   final Widget Function(T model) builder;
+//   final List<String>? fields; // optional field filter
+//
+//   const ReactiveBuilder({
+//     required this.model,
+//     required this.builder,
+//     this.fields,
+//     super.key,
+//   });
+//
+//   @override
+//   ReactiveBuilderState<T> createState() => ReactiveBuilderState<T>();
+// }
+//
+// class ReactiveBuilderState<T extends ReactiveModel>
+//     extends State<ReactiveBuilder<T>> {
+//   void _onChange() => setState(() {});
+//
+//   @override
+//   void initState() {
+//     super.initState();
+//     if (widget.fields != null) {
+//       for (var f in widget.fields!) {
+//         widget.model.addListener(_onChange, field: f);
+//       }
+//     } else {
+//       widget.model.addListener(_onChange); // listen to all fields
+//     }
+//   }
+//
+//   @override
+//   void dispose() {
+//     if (widget.fields != null) {
+//       for (var f in widget.fields!) {
+//         widget.model.removeListener(_onChange, field: f);
+//       }
+//     } else {
+//       widget.model.removeListener(_onChange);
+//     }
+//     super.dispose();
+//   }
+//
+//   @override
+//   Widget build(BuildContext context) => widget.builder(widget.model);
+// }
 
 ///Try Version
 // import 'package:flutter/material.dart';